# swagger.yaml
openapi: 3.0.0
info:
  title: Lab 5 - CRUD & Auth API
  description: API cho quản lý sản phẩm, nhà cung cấp và xác thực người dùng
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /users/register:
    post:
      summary: Đăng ký tài khoản
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: Đăng ký thành công
  /users/login:
    post:
      summary: Đăng nhập
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Đăng nhập thành công
  /users/forgot:
    post:
      summary: Quên mật khẩu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserForgot'
      responses:
        '200':
          description: Đổi mật khẩu thành công
  /users/logout:
    get:
      summary: Đăng xuất
      responses:
        '200':
          description: Đăng xuất thành công
  /products:
    get:
      summary: Lấy danh sách sản phẩm (hỗ trợ tìm kiếm, lọc)
      parameters:
        - name: search
          in: query
          description: Tìm theo tên sản phẩm (hỗ trợ tiếng Việt có dấu)
          required: false
          schema:
            type: string
        - name: supplier
          in: query
          description: Lọc theo id nhà cung cấp
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Danh sách sản phẩm
    post:
      summary: Thêm sản phẩm mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Thêm sản phẩm thành công
  /products/{id}:
    get:
      summary: Lấy chi tiết sản phẩm
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chi tiết sản phẩm
    put:
      summary: Cập nhật sản phẩm
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Cập nhật thành công
    delete:
      summary: Xóa sản phẩm
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Xóa thành công
  /suppliers:
    get:
      summary: Lấy danh sách nhà cung cấp
      responses:
        '200':
          description: Danh sách nhà cung cấp
    post:
      summary: Thêm nhà cung cấp mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Supplier'
      responses:
        '201':
          description: Thêm nhà cung cấp thành công
  /suppliers/{id}:
    get:
      summary: Lấy chi tiết nhà cung cấp
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chi tiết nhà cung cấp
    put:
      summary: Cập nhật nhà cung cấp
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Supplier'
      responses:
        '200':
          description: Cập nhật thành công
    delete:
      summary: Xóa nhà cung cấp
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Xóa thành công
components:
  schemas:
    Product:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        supplier:
          type: string
    Supplier:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
    UserRegister:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    UserLogin:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    UserForgot:
      type: object
      properties:
        username:
          type: string
        newPassword:
          type: string